Full Workable app, things that dont work: undo changes, delete object, validation

+++ src/main/java/rcas/model/AxleTireModel.java
@@ -5,11 +5,11 @@ import com.fasterxml.jackson.annotation.*;
 
 public class AxleTireModel {
     private String name;
-    private long slipAngleCoefficientC;
-    private long slipAngleCoefficientB;
-    private long slipAngleCoefficientE;
-    private long loadCoefficientKA;
-    private long loadCoefficientKB;
+    private double slipAngleCoefficientC;
+    private double slipAngleCoefficientB;
+    private double slipAngleCoefficientE;
+    private double loadCoefficientKA;
+    private double loadCoefficientKB;
 
     @JsonProperty("name")
     public String getName() { return name; }
@@ -17,27 +17,27 @@ public class AxleTireModel {
     public void setName(String value) { this.name = value; }
 
     @JsonProperty("slipAngleCoefficientC")
-    public long getSlipAngleCoefficientC() { return slipAngleCoefficientC; }
+    public double getSlipAngleCoefficientC() { return slipAngleCoefficientC; }
     @JsonProperty("slipAngleCoefficientC")
-    public void setSlipAngleCoefficientC(long value) { this.slipAngleCoefficientC = value; }
+    public void setSlipAngleCoefficientC(double value) { this.slipAngleCoefficientC = value; }
 
     @JsonProperty("slipAngleCoefficientB")
-    public long getSlipAngleCoefficientB() { return slipAngleCoefficientB; }
+    public double getSlipAngleCoefficientB() { return slipAngleCoefficientB; }
     @JsonProperty("slipAngleCoefficientB")
-    public void setSlipAngleCoefficientB(long value) { this.slipAngleCoefficientB = value; }
+    public void setSlipAngleCoefficientB(double value) { this.slipAngleCoefficientB = value; }
 
     @JsonProperty("slipAngleCoefficientE")
-    public long getSlipAngleCoefficientE() { return slipAngleCoefficientE; }
+    public double getSlipAngleCoefficientE() { return slipAngleCoefficientE; }
     @JsonProperty("slipAngleCoefficientE")
-    public void setSlipAngleCoefficientE(long value) { this.slipAngleCoefficientE = value; }
+    public void setSlipAngleCoefficientE(double value) { this.slipAngleCoefficientE = value; }
 
     @JsonProperty("loadCoefficientKA")
-    public long getLoadCoefficientKA() { return loadCoefficientKA; }
+    public double getLoadCoefficientKA() { return loadCoefficientKA; }
     @JsonProperty("loadCoefficientKA")
-    public void setLoadCoefficientKA(long value) { this.loadCoefficientKA = value; }
+    public void setLoadCoefficientKA(double value) { this.loadCoefficientKA = value; }
 
     @JsonProperty("loadCoefficientKB")
-    public long getLoadCoefficientKB() { return loadCoefficientKB; }
+    public double getLoadCoefficientKB() { return loadCoefficientKB; }
     @JsonProperty("loadCoefficientKB")
-    public void setLoadCoefficientKB(long value) { this.loadCoefficientKB = value; }
+    public void setLoadCoefficientKB(double value) { this.loadCoefficientKB = value; }
 }
